+++ app/src/main/kotlin/org/dbtools/sample/work/work/SimplePeriodicWorker.kt
@@ -2,20 +2,20 @@ package org.dbtools.sample.work.work
 
 import android.content.Context
 import androidx.annotation.WorkerThread
+import androidx.work.CoroutineWorker
 import androidx.work.Data
-import androidx.work.Worker
 import androidx.work.WorkerParameters
 import timber.log.Timber
 
-class SimplePeriodicWorker(context: Context, params: WorkerParameters) : Worker(context, params) {
+class SimplePeriodicWorker(context: Context, params: WorkerParameters) : CoroutineWorker(context, params) {
     @WorkerThread
-    override fun doWork(): Result {
+    override suspend fun doWork(): Payload {
         val inputText = inputData.getString(KEY_TEXT)
 
         logProgress("RUNNING PERIODIC: Text: [$inputText]")
 
         // return result
-        return Result.SUCCESS
+        return Payload(Result.SUCCESS)
     }
 
     private fun logProgress(progress: String) {