+++ app/src/main/kotlin/org/dbtools/sample/work/work/WorkScheduler.kt
@@ -1,6 +1,7 @@
 package org.dbtools.sample.work.work
 
 import androidx.work.Constraints
+import androidx.work.CoroutineWorker
 import androidx.work.Data
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.ExistingWorkPolicy
@@ -9,7 +10,6 @@ import androidx.work.OneTimeWorkRequest
 import androidx.work.PeriodicWorkRequest
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
-import androidx.work.Worker
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -62,7 +62,7 @@ class WorkScheduler
      * - Requires a Network Connection
      * - Requires charging
      */
-    private inline fun <reified T: Worker> createStandardPeriodicWorkRequest(repeatInterval: Long, timeUnit: TimeUnit, inputDataBuilder: Data.Builder = Data.Builder()): PeriodicWorkRequest {
+    private inline fun <reified T: CoroutineWorker> createStandardPeriodicWorkRequest(repeatInterval: Long, timeUnit: TimeUnit, inputDataBuilder: Data.Builder = Data.Builder()): PeriodicWorkRequest {
         val workerConstraintsBuilder = Constraints.Builder()
             .setRequiredNetworkType(NetworkType.CONNECTED)